* master:
  add valor e aliquotas de imposto pis e confis
  Adiciona campo observacoes ao model nfe
  Adiciona logger configurável
  ajuste na quantidade de item da nota para float

+++ lib/enotas_nfe/model/produto.rb
@@ -10,7 +10,7 @@ module EnotasNfe
       attribute :codigo, String
       attribute :descricao, String
       attribute :ncm, String
-      attribute :quantidade, Integer
+      attribute :quantidade, Float
       attribute :unidadeMedida, String
       attribute :valorUnitario, Float
       attribute :descontos, Float
